pyproject.toml
[tool.poetry]
name = "ragthoven"
version = "0.2.3"
description = ""
authors = ["Gregor Karetka <karetka.gregor@gmail.com>", "Marek Suppa <marek@mareksuppa.com>", "Demetris Skottis <demetris.s@hotmail.com>"]
readme = "README.md"
packages = [{ include = "ragthoven" }]

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.21.0"
pyyaml = "^6.0.2"
chromadb = "^0.5.5"
flashrank = "^0.2.9"
jinja2 = "^3.1.4"
openai = "^1.42.0"
litellm = "^1.44.14"
python-dotenv = "^0.19.2"
transformers = "^4.48"
sentence-transformers = "^3.1.1"
typer = "^0.12.5"
accelerate = "^1.8.0"
protobuf = "<3.20"
tqdm = "^4.66.5"
einops = "^0.8.0"

[tool.poetry.scripts]
ragthoven = "ragthoven.standalone:app"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.3"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
